feat(packages/sui-domain): Move to ESModules
