+++ packages/sui-domain/src/fetcher/factory.js
@@ -1,4 +1,4 @@
-import AxiosFetcher from './AxiosFetcher'
+import AxiosFetcher from './AxiosFetcher.js'
 
 export default class FetcherFactory {
   static httpFetcher = ({config}) => new AxiosFetcher({config})
