+++ packages/sui-domain/src/index.js
@@ -1,8 +1,8 @@
-export {default as Entity} from './Entity'
-export {default as EntryPointFactory} from './EntryPointFactory'
-export {default as FetcherFactory} from './fetcher/factory'
-export {default as Mapper} from './Mapper'
-export {default as Repository} from './Repository'
-export {default as Service} from './Service'
-export {default as UseCase} from './UseCase'
-export {default as ValueObject} from './ValueObject'
+export {default as Entity} from './Entity.js'
+export {default as EntryPointFactory} from './EntryPointFactory.js'
+export {default as FetcherFactory} from './fetcher/factory.js'
+export {default as Mapper} from './Mapper.js'
+export {default as Repository} from './Repository.js'
+export {default as Service} from './Service.js'
+export {default as UseCase} from './UseCase.js'
+export {default as ValueObject} from './ValueObject.js'
